+++ model/CHANGELOG.md
@@ -4,13 +4,14 @@ This file documents changes to the `workbench-model` library, including notes on
 
 ## 0.7
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-model" % "0.7-58a51e7"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-model" % "0.7-xxxxxxx"`
 
 ### Added
 
 - Model objects for pet service account keys
 - Support for base 64 encoding/decoding value objects
 - JSON formatters for java.time.Instant
+- emailOf function on WorkbenchSubject
 
 ## 0.6